bridge_service/requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
discord.py>=2.6,<3.0
python-multipart>=0.0.20,<1.0
